Decouple controller logic from UI #32
