+++ knes-compose-ui/src/main/kotlin/knes/compose/ComposeMain.kt
@@ -30,7 +30,6 @@ You should have received a copy of the GNU General Public License along with
 this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.focusable
 import androidx.compose.foundation.layout.*
@@ -44,133 +43,36 @@ import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.input.key.Key
-import androidx.compose.ui.input.key.KeyEventType
-import androidx.compose.ui.input.key.key
-import androidx.compose.ui.input.key.type
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
 import androidx.compose.ui.window.rememberWindowState
 import knes.controllers.KeyboardController
-import kotlinx.coroutines.delay
 import knes.emulator.NES
-import knes.emulator.input.InputHandler
 import knes.emulator.ui.GUIAdapter
-import knes.emulator.ui.ScreenView
-import java.awt.event.KeyEvent
+import kotlinx.coroutines.delay
 import javax.swing.JFileChooser
 import javax.swing.filechooser.FileNameExtensionFilter
-
-@Composable
-fun nesScreenRenderer(screenView: ComposeScreenView) {
-    var frameCount by remember { mutableStateOf(0) }
-    var currentBitmap by remember { mutableStateOf(screenView.getFrameBitmap()) }
-    val baseScale = screenView.scale
-    val isMacOS = System.getProperty("os.name").lowercase().contains("mac")
-    val scale = if (isMacOS) baseScale * 2 else baseScale
-
-    val scaledWidth = 512 * scale 
-    val scaledHeight = 480 * scale
-
-    DisposableEffect(Unit) {
-        screenView.onFrameReady = {
-            currentBitmap = screenView.getFrameBitmap()
-            frameCount++
-        }
-
-        onDispose {
-            screenView.onFrameReady = null
-        }
-    }
-
-    Canvas(
-        modifier = Modifier
-            .width(scaledWidth.dp)
-            .height(scaledHeight.dp)
-    ) {
-        drawImage(
-            image = currentBitmap,
-            dstSize = IntSize(scaledWidth, scaledHeight)
-        )
-    }
-}
 
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() = application {
     val windowState = rememberWindowState(width = 800.dp, height = 700.dp)
     var isEmulatorRunning by remember { mutableStateOf(false) }
 
-    val inputHandler = ComposeInputHandler()
+    val inputHandler = ComposeKeyboardInputHandler(KeyboardController())
 
     val screenView = remember { ComposeScreenView(1) }
-
     val nes = remember { NES(GUIAdapter(inputHandler, screenView)) }
     val composeUI = remember { ComposeUI(nes, screenView, inputHandler) }
-
-    fun mapKeyCode(key: Key): Int {
-        return when (key) {
-            Key.Z -> KeyEvent.VK_Z
-            Key.X -> KeyEvent.VK_X
-            Key.Spacebar -> KeyEvent.VK_ENTER
-            Key.V -> KeyEvent.VK_SPACE
-            Key.DirectionUp -> KeyEvent.VK_UP
-            Key.DirectionDown -> KeyEvent.VK_DOWN
-            Key.DirectionLeft -> KeyEvent.VK_LEFT
-            Key.DirectionRight -> KeyEvent.VK_RIGHT
-            else -> 0
-        }
-    }
-
-    // Define a function to map AWT key codes to their names
-    fun getKeyName(keyCode: Int): String {
-        return when (keyCode) {
-            KeyEvent.VK_Z -> "Z"
-            KeyEvent.VK_X -> "X"
-            KeyEvent.VK_ENTER -> "ENTER"
-            KeyEvent.VK_SPACE -> "SPACE"
-            KeyEvent.VK_UP -> "UP"
-            KeyEvent.VK_DOWN -> "DOWN"
-            KeyEvent.VK_LEFT -> "LEFT"
-            KeyEvent.VK_RIGHT -> "RIGHT"
-            else -> "UNKNOWN"
-        }
-    }
-
     val focusRequester = remember { FocusRequester() }
 
     Window(
         onCloseRequest = ::exitApplication,
         title = "kNES Emulator",
         state = windowState,
-        onKeyEvent = { event ->
-            val keyCode = if (event.key == Key.Enter) {
-                KeyEvent.VK_ENTER
-            } else {
-                mapKeyCode(event.key)
-            }
-
-            if (keyCode != 0) {
-                println("Key event: ${event.type} ${event.key} keyCode: $keyCode (${getKeyName(keyCode)})")
-
-                when (event.type) {
-                    KeyEventType.KeyDown -> {
-                        composeUI.inputHandler.setKeyState(keyCode, true)
-                        true
-                    }
-                    KeyEventType.KeyUp -> {
-                        composeUI.inputHandler.setKeyState(keyCode, false)
-                        true
-                    }
-                    else -> true  // Always consume key events
-                }
-            } else {
-                false
-            }
-        },
-        focusable = true  // Make the window focusable
+        onKeyEvent = inputHandler::keyEventHandler,
+        focusable = true
     ) {
         LaunchedEffect(Unit) {
             focusRequester.requestFocus()
@@ -178,21 +80,17 @@ fun main() = application {
 
         LaunchedEffect(Unit) {
             while (true) {
-                delay(1000) // Request focus every second
+                delay(1000)
                 focusRequester.requestFocus()
             }
         }
 
         MaterialTheme {
             Surface(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .focusRequester(focusRequester)
-                    .focusable()
+                modifier = Modifier.fillMaxSize().focusRequester(focusRequester).focusable()
             ) {
                 Column(
-                    modifier = Modifier.fillMaxSize().padding(16.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
+                    modifier = Modifier.fillMaxSize().padding(16.dp), horizontalAlignment = Alignment.CenterHorizontally
                 ) {
                     Text(
                         text = "kNES Emulator 🎮",
@@ -203,20 +101,11 @@ fun main() = application {
                         modifier = Modifier.fillMaxWidth().padding(top = 16.dp),
                         horizontalArrangement = Arrangement.SpaceEvenly
                     ) {
-                        Button(
-                            onClick = {
-                                if (isEmulatorRunning) {
-                                    composeUI.stopEmulator()
-                                } else {
-                                    composeUI.startEmulator()
-                                }
-                                isEmulatorRunning = !isEmulatorRunning
-                                // Request focus after button click
-                                focusRequester.requestFocus()
-                            }
-                        ) {
-                            Text(if (isEmulatorRunning) "Stop Emulator" else "Start Emulator")
-                        }
+                        Button(onClick = {
+                            if (isEmulatorRunning) composeUI.stopEmulator() else composeUI.startEmulator()
+                            isEmulatorRunning = !isEmulatorRunning
+                            focusRequester.requestFocus()
+                        }) { Text(if (isEmulatorRunning) "Stop Emulator" else "Start Emulator") }
 
                         Button(
                             onClick = {
@@ -232,8 +121,7 @@ fun main() = application {
                                     }
                                 }
                                 focusRequester.requestFocus()
-                            }
-                        ) {
+                            }) {
                             Text("Load ROM")
                         }
                     }
@@ -242,32 +130,30 @@ fun main() = application {
                         horizontalArrangement = Arrangement.SpaceEvenly
                     ) {
                         Box(
-                            modifier = Modifier.weight(1f),
-                            contentAlignment = Alignment.Center
+                            modifier = Modifier.weight(1f), contentAlignment = Alignment.Center
                         ) {
-                            nesScreenRenderer(screenView)
+                            composeUI.nesScreenRenderer()
                         }
                         Column {
-                        Box(
-                            modifier = Modifier.weight(1f),
-                            contentAlignment = Alignment.Center
-                        ) {
-                            Image(
-                                painter = painterResource("frame.png"),
-                                contentDescription = "NES Frame",
-                                modifier = Modifier.size(256.dp, 240.dp)
-                            )
+                            Box(
+                                modifier = Modifier.weight(1f), contentAlignment = Alignment.Center
+                            ) {
+                                Image(
+                                    painter = painterResource("frame.png"),
+                                    contentDescription = "NES Frame",
+                                    modifier = Modifier.size(256.dp, 240.dp)
+                                )
+                            }
+                            Box(
+                                modifier = Modifier.weight(1f), contentAlignment = Alignment.Center
+                            ) {
+                                Image(
+                                    painter = painterResource("logo.png"),
+                                    contentDescription = "NES Frame",
+                                    modifier = Modifier.size(256.dp, 240.dp)
+                                )
+                            }
                         }
-                        Box(
-                            modifier = Modifier.weight(1f),
-                            contentAlignment = Alignment.Center
-                        ) {
-                            Image(
-                                painter = painterResource("logo.png"),
-                                contentDescription = "NES Frame",
-                                modifier = Modifier.size(256.dp, 240.dp)
-                            )
-                        }}
 
                     }
                 }
@@ -275,3 +161,5 @@ fun main() = application {
         }
     }
 }
+
+
