+++ knes-controllers/src/main/kotlin/knes/controllers/KeyboardController.kt
@@ -13,51 +13,50 @@
 
 package knes.controllers
 
-import java.awt.event.KeyAdapter
+import knes.emulator.input.InputHandler
 import java.awt.event.KeyEvent
-import javax.swing.JComponent
+
 
 class KeyboardController : ControllerProvider {
-    private val keyStates = mutableMapOf<Int, Boolean>()
-    private val buttonMappings = mutableMapOf<ControllerProvider.NESButton, Int>().apply {
-        this[ControllerProvider.NESButton.A] = KeyEvent.VK_Z
-        this[ControllerProvider.NESButton.B] = KeyEvent.VK_X
-        this[ControllerProvider.NESButton.START] = KeyEvent.VK_ENTER
-        this[ControllerProvider.NESButton.SELECT] = KeyEvent.VK_SPACE
-        this[ControllerProvider.NESButton.UP] = KeyEvent.VK_UP
-        this[ControllerProvider.NESButton.DOWN] = KeyEvent.VK_DOWN
-        this[ControllerProvider.NESButton.LEFT] = KeyEvent.VK_LEFT
-        this[ControllerProvider.NESButton.RIGHT] = KeyEvent.VK_RIGHT
+    private val keyStates = ShortArray(InputHandler.Companion.NUM_KEYS) { 0 }
+    private val keyMapping = IntArray(InputHandler.Companion.NUM_KEYS) { 0 }
+
+    init {
+        keyMapping[InputHandler.KEY_A] = KeyEvent.VK_Z
+        keyMapping[InputHandler.KEY_B] = KeyEvent.VK_X
+        keyMapping[InputHandler.KEY_START] = KeyEvent.VK_ENTER
+        keyMapping[InputHandler.KEY_SELECT] = KeyEvent.VK_SPACE
+        keyMapping[InputHandler.KEY_UP] = KeyEvent.VK_UP
+        keyMapping[InputHandler.KEY_DOWN] = KeyEvent.VK_DOWN
+        keyMapping[InputHandler.KEY_LEFT] = KeyEvent.VK_LEFT
+        keyMapping[InputHandler.KEY_RIGHT] = KeyEvent.VK_RIGHT
     }
 
-    override fun getButtonState(button: ControllerProvider.NESButton): Short {
-        val keyCode = buttonMappings[button]
-        // NES expects 0x41 for pressed, 0x40 for not pressed
-        return if (keyCode != null && keyStates[keyCode] == true) 0x41 else 0x40
+    override fun setKeyState(keyCode: Int, isPressed: Boolean) {
+        for (i in keyMapping.indices) {
+            if (keyMapping[i] == keyCode) {
+                keyStates[i] = if (isPressed) 0x41 else 0x40
+            }
+        }
     }
 
-    override fun mapButton(button: ControllerProvider.NESButton, code: Int) {
-        buttonMappings[button] = code
+    override fun getKeyState(padKey: Int): Short {
+        return keyStates[padKey]
     }
 
-    override fun update() {
-        // No need to update key states here, handled by key adapter
-    }
-
-    fun registerKeyComponent(component: JComponent) {
-        component.addKeyListener(object : KeyAdapter() {
-            override fun keyPressed(e: KeyEvent) { handleKeyEvent(e.keyCode, true) }
-            override fun keyReleased(e: KeyEvent) { handleKeyEvent(e.keyCode, false) }
-        })
-        component.isFocusable = true
-        component.requestFocus()
-    }
-
-    private fun handleKeyEvent(keyCode: Int, pressed: Boolean) {
-        keyStates[keyCode] = pressed
-    }
-
-    fun setKeyState(keyCode: Int, isPressed: Boolean) {
-        keyStates[keyCode] = isPressed
+    companion object {
+        fun getKeyName(keyCode: Int): String {
+            return when (keyCode) {
+                KeyEvent.VK_Z -> "Z"
+                KeyEvent.VK_X -> "X"
+                KeyEvent.VK_ENTER -> "ENTER"
+                KeyEvent.VK_SPACE -> "SPACE"
+                KeyEvent.VK_UP -> "UP"
+                KeyEvent.VK_DOWN -> "DOWN"
+                KeyEvent.VK_LEFT -> "LEFT"
+                KeyEvent.VK_RIGHT -> "RIGHT"
+                else -> "UNKNOWN"
+            }
+        }
     }
 }
