+++ knes-emulator/build.gradle
@@ -11,6 +11,10 @@
  *
  */
 
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+
+
+
 plugins {
     id 'java'
     id 'org.jetbrains.kotlin.jvm'
@@ -21,7 +25,6 @@ repositories {
 }
 
 dependencies {
-    implementation project(':knes-controllers')
     testImplementation 'junit:junit:4.13.2'
 }
 
@@ -29,7 +32,7 @@ kotlin {
     jvmToolchain(11)
 }
 
-tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).configureEach {
+tasks.withType(KotlinCompile).configureEach {
     kotlinOptions {
         jvmTarget = '11'
         apiVersion = '2.2'
